# backend/src/controllers/auth.controller.ts
import { Request, Response, NextFunction } from 'express'
import * as authService from '../services/auth.service'

/**
 * @swagger
 * tags:
 *   name: Auth
 *   description: Authentication endpoints
 */

/**
 * @swagger
 * /api/auth/register:
 *   post:
 *     summary: Register a new user
 *     tags: [Auth]
 *     security: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [name, email, password, department]
 *             properties:
 *               name:
 *                 type: string
 *               email:
 *                 type: string
 *                 format: email
 *               password:
 *                 type: string
 *                 minLength: 6
 *               department:
 *                 type: string
 *     responses:
 *       201:
 *         description: User created
 *       409:
 *         description: Email already in use
 *       422:
 *         $ref: '#/components/schemas/Error'
 */
export async function register(req: Request, res: Response, next: NextFunction): Promise<void> {
  try {
    const result = await authService.register(req.body)
    res.status(201).json(result)
  } catch (err) {
    next(err)
  }
}

/**
 * @swagger
 * /api/auth/login:
 *   post:
 *     summary: Login and receive a JWT token
 *     tags: [Auth]
 *     security: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email, password]
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *               password:
 *                 type: string
 *     responses:
 *       200:
 *         description: Login successful, returns token
 *       401:
 *         description: Invalid credentials
 */
export async function login(req: Request, res: Response, next: NextFunction): Promise<void> {
  try {
    const result = await authService.login(req.body)
    res.json(result)
  } catch (err) {
    next(err)
  }
}

/**
 * @swagger
 * /api/auth/me:
 *   get:
 *     summary: Get current authenticated user info
 *     tags: [Auth]
 *     responses:
 *       200:
 *         description: Current user payload from JWT
 *       401:
 *         description: Unauthorized
 */
export function me(req: Request, res: Response): void {
  res.json({ user: req.user })
}

/**
 * @swagger
 * /api/auth/profile:
 *   get:
 *     summary: Get current user full profile
 *     tags: [Auth]
 *     responses:
 *       200:
 *         description: Full user profile from database
 *       401:
 *         description: Unauthorized
 */
export async function profile(req: Request, res: Response, next: NextFunction): Promise<void> {
  try {
    const user = await authService.getProfile(req.user!.sub)
    res.json(user)
  } catch (err) {
    next(err)
  }
}

/**
 * @swagger
 * /api/auth/profile:
 *   patch:
 *     summary: Update the authenticated user's own profile (name, department, avatar)
 *     tags: [Auth]
 *     responses:
 *       200:
 *         description: Updated profile
 */
export async function updateProfile(req: Request, res: Response, next: NextFunction): Promise<void> {
  try {
    res.json(await authService.updateProfile(req.user!.sub, req.body))
  } catch (err) {
    next(err)
  }
}

/**
 * @swagger
 * /api/auth/password:
 *   patch:
 *     summary: Change current user password
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [currentPassword, newPassword]
 *             properties:
 *               currentPassword:
 *                 type: string
 *               newPassword:
 *                 type: string
 *                 minLength: 6
 *     responses:
 *       204:
 *         description: Password changed
 *       401:
 *         description: Current password incorrect
 */
export async function changePassword(
  req: Request,
  res: Response,
  next: NextFunction,
): Promise<void> {
  try {
    const { currentPassword, newPassword } = req.body
    await authService.changePassword(req.user!.sub, currentPassword, newPassword)
    res.status(204).send()
  } catch (err) {
    next(err)
  }
}