# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockIQ — WMS API",
    "version": "1.0.0",
    "description": "API REST do StockIQ (WMS / ERP Lite).\n\n**Autenticação:** faça `POST /api/auth/login`, copie o `token` retornado e clique em **Authorize** (esquema `bearerAuth`).\n\n**Perfis:** `admin` > `supervisor` > `operator` > `viewer`. Cada operação lista os perfis permitidos.\n\n**Erros:** respostas de erro seguem `{ \"message\": string }`; validações retornam 422 com `errors[]`."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "name": {
            "type": "string",
            "example": "Carlos Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@stockiq.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "supervisor",
              "operator",
              "viewer"
            ],
            "example": "admin"
          },
          "department": {
            "type": "string",
            "example": "TI"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "pending"
            ],
            "example": "active"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AuthResult": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "JwtPayload": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "sub": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "email": {
                "type": "string",
                "example": "admin@stockiq.com"
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "supervisor",
                  "operator",
                  "viewer"
                ],
                "example": "admin"
              },
              "iat": {
                "type": "integer"
              },
              "exp": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "Movement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "type": {
            "type": "string",
            "enum": [
              "entry",
              "exit",
              "transfer",
              "loss",
              "adjustment",
              "inventory"
            ],
            "example": "entry"
          },
          "quantity": {
            "type": "integer",
            "example": 50
          },
          "unitCost": {
            "type": "number",
            "example": 22.5
          },
          "totalValue": {
            "type": "number",
            "example": 1125
          },
          "invoiceNumber": {
            "type": "string",
            "nullable": true
          },
          "lotNumber": {
            "type": "string",
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "exitReason": {
            "type": "string",
            "enum": [
              "sale",
              "transfer",
              "loss",
              "break",
              "internal"
            ],
            "example": "sale",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "productId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "supplierId": {
            "type": "string",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "fromAddressId": {
            "type": "string",
            "nullable": true
          },
          "toAddressId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string"
              },
              "internalCode": {
                "type": "string"
              }
            }
          },
          "supplier": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string",
                "example": "Nome"
              }
            },
            "nullable": true
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string",
                "example": "Nome"
              }
            }
          },
          "fromAddress": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "code": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "toAddress": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "code": {
                "type": "string"
              }
            },
            "nullable": true
          }
        }
      },
...
      "DashboardSummary": {
        "type": "object",
        "properties": {
          "products": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 26
              },
              "active": {
                "type": "integer",
                "example": 25
              },
              "lowStock": {
                "type": "integer",
                "example": 5
              },
              "outStock": {
                "type": "integer",
                "example": 1
              }
            }
          },
          "movements": {
            "type": "object",
            "properties": {
              "today": {
                "type": "integer",
                "example": 6
              },
              "todayEntries": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 2
                  },
                  "value": {
                    "type": "number",
                    "example": 1350
                  }
                }
              },
              "todayExits": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 4
                  },
                  "value": {
                    "type": "number",
                    "example": 2800
                  }
                }
              }
            }
          },
          "users": {
            "type": "object",
            "properties": {
              "active": {
                "type": "integer",
                "example": 6
              }
            }
          },
          "lots": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 10
              },
              "expiringSoon": {
                "type": "integer",
                "example": 3
              },
              "validPercentage": {
                "type": "integer",
                "example": 60
              }
            }
          },
          "inventory": {
            "type": "object",
            "properties": {
              "pendingCounts": {
                "type": "integer",
                "example": 3
              },
              "accuracy": {
                "type": "number",
                "nullable": true,
                "example": 99.2
              }
            }
          },
          "warehouse": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 64
              },
              "free": {
                "type": "integer",
                "example": 30
              },
              "occupied": {
                "type": "integer",
                "example": 27
              },
              "blocked": {
                "type": "integer",
                "example": 4
              },
              "occupancyRate": {
                "type": "integer",
                "example": 42
              }
            }
          },
          "stock": {
            "type": "object",
            "properties": {
              "totalPurchaseValue": {
                "type": "number",
                "example": 182000.5
              },
              "totalSaleValue": {
                "type": "number",
                "example": 341000
              }
            }
          },
          "recentMovements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movement"
            }
          }
        }
      },
      "MovementTrendPoint": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "out. de 26"
          },
          "entries": {
            "type": "integer",
            "example": 900
          },
          "exits": {
            "type": "integer",
            "example": 850
          },
          "balance": {
            "type": "integer",
            "example": 50
          }
        }
      },
      "CategoryDistributionItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "name": {
            "type": "string",
            "example": "Eletrônicos"
          },
          "color": {
            "type": "string",
            "example": "#2563eb"
          },
          "productCount": {
            "type": "integer",
            "example": 8
          },
          "percentage": {
            "type": "integer",
            "example": 30
          },
          "stockValue": {
            "type": "number",
            "example": 52000
          }
        }
      },
      "TopProduct": {
        "type": "object",
        "properties": {
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string"
              },
              "internalCode": {
                "type": "string"
              },
              "currentStock": {
                "type": "integer"
              },
              "salePrice": {
                "type": "number"
              }
            }
          },
          "movementCount": {
            "type": "integer",
            "example": 54
          },
          "totalQuantity": {
            "type": "integer",
            "example": 980
          },
          "totalValue": {
            "type": "number",
            "example": 48000
          },
          "trend": {
            "type": "integer",
            "example": 12,
            "description": "% change in movement count, last 30 days vs previous 30."
          }
        }
      },
      "HeatmapCell": {
        "type": "object",
        "properties": {
          "day": {
            "type": "integer",
            "minimum": 0,
            "maximum": 6,
            "example": 2
          },
          "hour": {
            "type": "integer",
            "example": 14
          },
          "value": {
            "type": "integer",
            "example": 7
          }
        }
      },
...
      "AbcCurve": {
        "type": "object",
        "properties": {
          "A": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbcItem"
            }
          },
          "B": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbcItem"
            }
          },
          "C": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbcItem"
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "totalValue": {
                "type": "number",
                "example": 480000
              },
              "A": {
                "$ref": "#/components/schemas/AbcClassStat"
              },
              "B": {
                "$ref": "#/components/schemas/AbcClassStat"
              },
              "C": {
                "$ref": "#/components/schemas/AbcClassStat"
              }
            }
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Maria Souza"
          },
          "department": {
            "type": "string",
            "example": "Logística"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflito (valor único já existente ou registro vinculado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erro de validação do corpo/parâmetros",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Limite de requisições excedido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "summary": "Get dashboard summary (KPIs)",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Summary with products, movements, users, warehouse stats and stock value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/dashboard/movement-trend": {
      "get": {
        "summary": "Monthly entries vs exits trend",
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "schema": {
              "type": "integer",
              "default": 12
            },
            "description": "How many months to look back"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of monthly { month, entries, exits, balance }",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovementTrendPoint"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/dashboard/category-distribution": {
      "get": {
        "summary": "Product count and stock value by category",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Categories with product count, percentage and stock value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDistributionItem"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/dashboard/top-products": {
      "get": {
        "summary": "Top products by movement frequency",
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products ranked by number of movements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopProduct"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/dashboard/abc": {
      "get": {
        "summary": "ABC curve analysis (A=80%, B=15%, C=5% of value)",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Products classified into A, B and C by revenue contribution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbcCurve"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/dashboard/heatmap": {
      "get": {
        "summary": "Movement volume by weekday and hour",
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 90
            },
            "description": "How many days back to aggregate"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of { day (0=Sunday), hour, value }",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapCell"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "department"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  },
                  "department": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login and receive a JWT token",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns token"
          },
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get current authenticated user info",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Current user payload from JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtPayload"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "Get current user full profile",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Full user profile from database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "summary": "Update the authenticated user's own profile (name, department, avatar)",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/password": {
      "patch": {
        "summary": "Change current user password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    }
  }
}